requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=5.0.1=py310haa95532_0
anyio=4.2.0=py310haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h2bbff1b_0
arrow-cpp=16.1.0=h7cd61ee_0
asttokens=2.0.5=pyhd3eb1b0_0
async-lru=2.0.4=py310haa95532_0
attrs=23.1.0=py310haa95532_0
aws-c-auth=0.6.19=h2bbff1b_0
aws-c-cal=0.5.20=h2bbff1b_0
aws-c-common=0.8.5=h2bbff1b_0
aws-c-compression=0.2.16=h2bbff1b_0
aws-c-event-stream=0.2.15=hd77b12b_0
aws-c-http=0.6.25=h2bbff1b_0
aws-c-io=0.13.10=h2bbff1b_0
aws-c-mqtt=0.7.13=h2bbff1b_0
aws-c-s3=0.1.51=h2bbff1b_0
aws-c-sdkutils=0.1.6=h2bbff1b_0
aws-checksums=0.1.13=h2bbff1b_0
aws-crt-cpp=0.18.16=hd77b12b_0
aws-sdk-cpp=1.10.55=hd77b12b_0
babel=2.11.0=py310haa95532_0
beautifulsoup4=4.12.3=py310haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.6.2=py310haa95532_0
blosc=1.21.3=h6c2663c_0
boost-cpp=1.82.0=h59b6b97_2
bottleneck=1.3.7=py310h9128911_0
branca=0.6.0=py310haa95532_0
brotli=1.0.9=h2bbff1b_8
brotli-bin=1.0.9=h2bbff1b_8
brotli-python=1.0.9=py310hd77b12b_8
bzip2=1.0.8=h2bbff1b_6
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2024.9.24=haa95532_0
cachetools=5.3.3=py310haa95532_0
cairo=1.16.0=haedb8bc_5
certifi=2024.8.30=py310haa95532_0
cffi=1.17.1=py310h827c3e9_0
cfitsio=3.470=h2bbff1b_7
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py310haa95532_0
click-plugins=1.1.1=pyhd3eb1b0_0
cligj=0.7.2=pyhd3eb1b0_0
colorama=0.4.6=py310haa95532_0
comm=0.2.1=py310haa95532_0
contourpy=1.2.0=py310h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
exceptiongroup=1.2.0=py310haa95532_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.6.3=h5da7b33_0
fiona=1.9.5=py310h4ed8f06_0
folium=0.14.0=py310haa95532_0
fontconfig=2.14.1=hb33846d_3
fonttools=4.51.0=py310h2bbff1b_0
freetype=2.12.1=ha860e81_0
freexl=2.0.0=hd7a5696_0
gdal=3.6.2=py310h7670e6c_5
geodatasets=2024.8.0=pyhd8ed1ab_0
geographiclib=2.0=py310haa95532_0
geopandas=0.14.2=py310haa95532_0
geopandas-base=0.14.2=py310haa95532_0
geopy=2.4.1=py310haa95532_0
geos=3.8.0=h33f27b4_0
geotiff=1.7.0=h4545760_3
gflags=2.2.2=hd77b12b_1
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.43=py310haa95532_0
glib=2.78.4=hd77b12b_0
glib-tools=2.78.4=hd77b12b_0
glog=0.5.0=hd77b12b_1
h11=0.14.0=py310haa95532_0
hdf4=4.2.13=h712560f_2
hdf5=1.12.1=h51c971a_3
httpcore=1.0.2=py310haa95532_0
httpx=0.27.0=py310haa95532_0
icc_rt=2022.1.0=h6049295_2
idna=3.7=py310haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.28.0=py310haa95532_0
ipython=8.27.0=py310haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.8.1=py310haa95532_0
jedi=0.19.1=py310haa95532_0
jinja2=3.1.4=py310haa95532_0
joblib=1.4.2=py310haa95532_0
jpeg=9e=h827c3e9_3
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=py310haa95532_0
jsonschema-specifications=2023.7.1=py310haa95532_0
jupyter-lsp=2.2.0=py310haa95532_0
jupyter_client=8.6.0=py310haa95532_0
jupyter_core=5.7.2=py310haa95532_0
jupyter_events=0.10.0=py310haa95532_0
jupyter_server=2.14.1=py310haa95532_0
jupyter_server_terminals=0.4.4=py310haa95532_1
jupyterlab=4.2.5=py310haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.27.3=py310haa95532_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kealib=1.5.0=hde4a422_1
kiwisolver=1.4.4=py310hd77b12b_0
krb5=1.20.1=h5b6d351_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libabseil=20240116.2=cxx17_h5da7b33_0
libboost=1.82.0=h3399ecb_2
libbrotlicommon=1.0.9=h2bbff1b_8
libbrotlidec=1.0.9=h2bbff1b_8
libbrotlienc=1.0.9=h2bbff1b_8
libcurl=8.9.1=h0416ee5_0
libdeflate=1.17=h2bbff1b_1
libevent=2.1.12=h56d1f94_1
libffi=3.4.4=hd77b12b_1
libgdal=3.6.2=h0e70117_5
libglib=2.78.4=ha17d25a_0
libgrpc=1.62.2=hf25190f_0
libiconv=1.16=h2bbff1b_3
libkml=1.3.0=h63940dd_7
libnetcdf=4.8.1=h6685c40_4
libpng=1.6.39=h8cc25b3_0
libpq=12.17=h906ac69_0
libprotobuf=4.25.3=hf2fb9eb_0
libsodium=1.0.18=h62dcd97_0
libspatialindex=1.9.3=h6c2663c_0
libspatialite=5.1.0=h25d3e1c_1
libssh2=1.11.0=h291bd65_0
libthrift=0.15.0=h4364b78_2
libtiff=4.5.1=hd77b12b_0
libwebp-base=1.3.2=h2bbff1b_0
libxml2=2.13.1=h24da03e_2
libzip=1.8.0=h289538f_1
lz4-c=1.9.4=h2bbff1b_1
mapclassify=2.5.0=py310haa95532_0
markdown-it-py=2.2.0=py310haa95532_1
markupsafe=2.1.3=py310h2bbff1b_0
matplotlib-base=3.9.2=py310he19b0ae_0
matplotlib-inline=0.1.6=py310haa95532_0
mdurl=0.1.0=py310haa95532_0
minizip=4.0.3=hb68bac4_0
mistune=2.0.4=py310haa95532_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.10=py310h827c3e9_0
mkl_random=1.2.7=py310hc64d2fc_0
nbclient=0.8.0=py310haa95532_0
nbconvert=7.10.0=py310haa95532_0
nbformat=5.9.2=py310haa95532_0
nest-asyncio=1.6.0=py310haa95532_0
networkx=3.3=py310haa95532_0
notebook=7.2.2=py310haa95532_1
notebook-shim=0.2.3=py310haa95532_0
numexpr=2.8.7=py310h2cd9be0_0
numpy=1.26.4=py310h055cbcc_0
numpy-base=1.26.4=py310h65a83cf_0
openjpeg=2.5.2=hae555c5_0
openssl=3.3.2=h2466b09_0
orc=2.0.1=hd8d391b_0
overrides=7.4.0=py310haa95532_0
packaging=24.1=py310haa95532_0
pandas=2.2.2=py310h5da7b33_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre2=10.42=h0ff8eda_1
pillow=10.4.0=py310h827c3e9_0
pip=24.2=py310haa95532_0
pixman=0.40.0=h2bbff1b_1
platformdirs=3.10.0=py310haa95532_0
pooch=1.8.2=pyhd8ed1ab_0
poppler=22.12.0=h0bf3bde_3
poppler-data=0.4.11=haa95532_1
proj=9.3.1=ha107b6e_0
prometheus_client=0.14.1=py310haa95532_0
prompt-toolkit=3.0.43=py310haa95532_0
prompt_toolkit=3.0.43=hd3eb1b0_0
protobuf=4.25.3=py310h958608f_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=16.1.0=py310hc64d2fc_0
pybind11-abi=5=hd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.8.0=py310haa95532_2
pygments=2.15.1=py310haa95532_1
pyparsing=3.1.2=py310haa95532_0
pyproj=3.6.1=py310h8d66fa9_0
pysocks=1.7.1=py310haa95532_0
python=3.10.15=h4607a30_1
python-dateutil=2.9.0post0=py310haa95532_2
python-fastjsonschema=2.16.2=py310haa95532_0
python-json-logger=2.0.7=py310haa95532_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=py310haa95532_0
pywin32=305=py310h2bbff1b_0
pywinpty=2.0.10=py310h5da7b33_0
pyyaml=6.0.1=py310h2bbff1b_0
pyzmq=25.1.2=py310hd77b12b_0
qhull=2020.2=h59b6b97_2
re2=2022.04.01=hd77b12b_0
referencing=0.30.2=py310haa95532_0
requests=2.32.3=py310haa95532_0
rfc3339-validator=0.1.4=py310haa95532_0
rfc3986-validator=0.1.1=py310haa95532_0
rich=13.7.1=py310haa95532_0
rpds-py=0.10.6=py310h062c2fa_0
rtree=1.0.1=py310h2eaa2aa_0
scikit-learn=1.5.1=py310hc64d2fc_0
scipy=1.13.1=py310h8640f81_0
seaborn=0.13.2=py310haa95532_0
send2trash=1.8.2=py310haa95532_0
setuptools=75.1.0=py310haa95532_0
shapely=2.0.5=py310hd1cf0e4_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.2.1=hcdb6601_0
sniffio=1.3.0=py310haa95532_0
soupsieve=2.5=py310haa95532_0
sqlite=3.45.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.37.1=py310haa95532_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.3=py310haa95532_0
terminado=0.17.1=py310haa95532_0
threadpoolctl=3.5.0=py310h9909e9c_0
tiledb=2.3.3=hd8964de_3
tinycss2=1.2.1=py310haa95532_0
tk=8.6.14=h0416ee5_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py310haa95532_0
toolz=0.12.0=py310haa95532_0
tornado=6.4.1=py310h827c3e9_0
traitlets=5.14.3=py310haa95532_0
typing-extensions=4.11.0=py310haa95532_0
typing_extensions=4.11.0=py310haa95532_0
tzdata=2024a=h04d1e81_0
ucrt=10.0.22621.0=h57928b3_0
unicodedata2=15.1.0=py310h2bbff1b_0
uriparser=0.9.7=h2bbff1b_0
urllib3=2.2.3=py310haa95532_0
utf8proc=2.6.1=h2bbff1b_1
vc=14.40=h2eaa2aa_1
vc14_runtime=14.40.33810=ha82c5b3_21
vs2015_runtime=14.40.33810=h3bf8584_21
watchdog=4.0.1=py310haa95532_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310haa95532_1
websocket-client=1.8.0=py310haa95532_0
wheel=0.44.0=py310haa95532_0
widgetsnbextension=3.6.6=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
winpty=0.4.3=4
xerces-c=3.2.4=hd77b12b_1
xyzservices=2022.9.0=py310haa95532_1
xz=5.4.6=h8cc25b3_1
yaml=0.2.5=he774522_0
zeromq=4.3.5=hd77b12b_0
zlib=1.2.13=h8cc25b3_1
zstd=1.5.5=hd43e919_2
